+++ src/sw/allotmentclub/__init__.py
@@ -5,7 +5,8 @@ import warnings
 
 # VALUE_PER_MEMBER = 600000  # until end of 2015
 # VALUE_PER_MEMBER = 750000  # until end of 2016
-VALUE_PER_MEMBER = 650000
+# VALUE_PER_MEMBER = 650000  # until end of 2018
+VALUE_PER_MEMBER = 750000
 
 SCAN_IGNORE_TESTS = [re.compile(r'\.testing$').search,
                      re.compile(r'\.conftest$').search,